\docType{class}
\name{McmcdbSqlite}
\alias{McmcdbSqlite}
\title{\code{McmcdbSqlite} Class}
\description{
  SQLite backend for storing \code{mcmcdb} objects.
}
\section{Slots}{
  \description{
  \item{\code{connection}}{\code{"SQLiteConnection"}.
  Connection to the SQLite Database used to store the MCMC
  data.} }
}

